% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_addSimpleFeatures.R
\name{addSimpleFeatures}
\alias{addSimpleFeatures}
\title{Add Simple Features}
\usage{
addSimpleFeatures(map, sf, ...)
}
\arguments{
\item{map}{\code{\link[leaflet:leaflet]{leaflet::leaflet()}} Map object.}

\item{sf}{\code{\link[sf:sf]{sf::st_sf()}} object.}
}
\value{
An updated \code{\link[leaflet:leaflet]{leaflet::leaflet()}} map object.
}
\description{
Add simple features data to a \code{\link[leaflet:leaflet]{leaflet::leaflet()}} map.
}
