% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_verification.R
\name{extract_valid_names}
\alias{extract_valid_names}
\title{Extract valid names}
\usage{
extract_valid_names(x)
}
\arguments{
\item{x}{\code{character} \code{vector} object.}
}
\value{
\code{character} value.
}
\description{
Extract valid names in a vector.
}
