% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_verification.R
\name{n_valid_names}
\alias{n_valid_names}
\title{Number of valid names}
\usage{
n_valid_names(x)
}
\arguments{
\item{x}{\code{character} \code{vector} object.}
}
\value{
\code{logical} value.
}
\description{
Identify number of valid names in a vector.
}
