% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_render_data_validation_alerts.R
\name{render_data_validation_alerts}
\alias{render_data_validation_alerts}
\title{Render data validation alerts}
\usage{
render_data_validation_alerts(
  session,
  anchorId,
  alertIdPrefix,
  validation_results
)
}
\arguments{
\item{session}{Shiny session.}

\item{anchorId}{\code{character} Value indicating alert anchor identifier
(used with \code{\link[shinyBS:createAlert]{shinyBS::createAlert()}} and \code{\link[shinyBS:closeAlert]{shinyBS::closeAlert()}}.}

\item{alertIdPrefix}{\code{character} Value indicating prefixes for individual
alerts.}

\item{validation_results}{\code{list} Object with validation results
(produced using \code{\link[=validate_data]{validate_data()}}).}
}
\value{
An invisible \code{TRUE} value.
}
\description{
This function displays alerts related to invalid user inputs.
}
\details{
This function is used for the side-effect of creating alerts.
}
