% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sanitize_data.R
\name{sanitize_data}
\alias{sanitize_data}
\title{Sanitize data}
\usage{
sanitize_data(x)
}
\arguments{
\item{x}{\code{data.frame} Object.}
}
\value{
A \code{data.frame} object.
}
\description{
Sanitize data input by a user.
}
\details{
At the moment, this just replaces
zero-length character values with missing (\code{NA}) values.
}
