% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_site_leaflet_map.R
\name{site_leaflet_map}
\alias{site_leaflet_map}
\title{Initialize a leaflet map for adding site data.}
\usage{
site_leaflet_map()
}
\value{
A \code{\link[leaflet:leaflet]{leaflet::leaflet()}} object.
}
\description{
Initialize a leaflet map for adding site data.
}
