% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_nav_buttons.R
\name{update_nav_buttons}
\alias{update_nav_buttons}
\title{Update navigation buttons}
\usage{
update_nav_buttons(x)
}
\arguments{
\item{x}{\code{logical} Indicating if buttons should be enabled, or disabled.}
}
\value{
Invisible \code{TRUE}.
}
\description{
Enable or disable navigation buttons
}
