% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_verification.R
\name{valid_nchar}
\alias{valid_nchar}
\title{Valid number of characters}
\usage{
valid_nchar(x, length)
}
\arguments{
\item{x}{\code{character} \code{vector} object.}

\item{length}{\code{integer} maximum number of characters permitted.}
}
\value{
\code{logical} value.
}
\description{
Check that the valid names in a vector have an acceptable number of
characters.
}
