% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_validate_data.R
\name{validate_data}
\alias{validate_data}
\title{Validate data}
\usage{
validate_data(x)
}
\arguments{
\item{x}{\code{data.frame} Table object.}
}
\value{
A \code{list} object containing a series of \code{list} objects that contain
details for various validation checks. Each of the nested \code{list}
objects contains the following elements:

\describe{
\item{id}{A \code{character} value containing a unique identifier for the check.}
\item{title}{A \code{character} value containing a title for the check.}
\item{details}{A \code{character} value containing an explanation of the check.}
\item{success}{A \code{logical} value indicating if the check was passed or not.}
}
}
\description{
Validate data input by a user.
}
