% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{whattemplatemaker}
\alias{whattemplatemaker}
\title{whattemplatemaker: Interactive data template application for the What To Do application}
\description{
The What Template Maker is a tool that creates data templates for the What
To Do application. Since the What To Do application requires data in a very
specific format, this application is designed to help with preparing input
data for the What To Do application. After providing the names of sites,
features, and actions, a template Excel Spreadsheet can then be downloaded
from this application. The template Excel Spreadsheet can then be populated
with data (e.g. longitude/latitude of sites) and, subsequently, uploaded to
the What To Do application.
}
\examples{
if (interactive()) {
run_app()
}
}
